Rename key-manager to maas-api: go package renaming
