+++ key-manager/internal/handlers/keys.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/gin-gonic/gin"
 
-	"github.com/opendatahub-io/maas-billing/key-manager/internal/keys"
-	"github.com/opendatahub-io/maas-billing/key-manager/internal/teams"
+	"github.com/opendatahub-io/maas-billing/maas-api/internal/keys"
+	"github.com/opendatahub-io/maas-billing/maas-api/internal/teams"
 )
 
 // KeysHandler handles key-related endpoints
