+++ key-manager/internal/handlers/models.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/gin-gonic/gin"
 
-	"github.com/opendatahub-io/maas-billing/key-manager/internal/models"
+	"github.com/opendatahub-io/maas-billing/maas-api/internal/models"
 )
 
 // ModelsHandler handles model-related endpoints
