+++ key-manager/internal/keys/manager.go
@@ -14,7 +14,7 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
 
-	"github.com/opendatahub-io/maas-billing/key-manager/internal/teams"
+	"github.com/opendatahub-io/maas-billing/maas-api/internal/teams"
 )
 
 // Manager handles API key operations
