+++ key-manager/internal/usage/collector.go
@@ -14,7 +14,7 @@ import (
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 
-	"github.com/opendatahub-io/maas-billing/key-manager/internal/types"
+	"github.com/opendatahub-io/maas-billing/maas-api/internal/types"
 )
 
 // Collector handles usage data collection from Istio Prometheus metrics
